feat: extracting widget and refactoring
